+++ src/lib/components/tx/modal/ButtonSection.tsx
@@ -1,7 +1,12 @@
 import { Button } from "@chakra-ui/react";
 import { useRouter } from "next/router";
 
-import { useInternalNavigate, useCelatoneApp } from "lib/app-provider";
+import {
+  useInternalNavigate,
+  useCelatoneApp,
+  useLCDEndpoint,
+} from "lib/app-provider";
+import { getNavigationUrl } from "lib/components/ExplorerLink";
 import { CustomIcon } from "lib/components/icon";
 import { useOpenTxTab } from "lib/hooks";
 import type { ActionVariant, TxReceipt } from "lib/types";
@@ -23,10 +28,9 @@ export const ButtonSection = ({
   const navigate = useInternalNavigate();
   const openTxTab = useOpenTxTab("tx-page");
   const {
-    chainConfig: {
-      explorerLink: { proposal: explorerProposal },
-    },
+    chainConfig: { explorerLink },
   } = useCelatoneApp();
+  const lcdEndpoint = useLCDEndpoint();
 
   const openTxExplorer = () => {
     const txHash = receipts
@@ -40,7 +44,15 @@ export const ButtonSection = ({
     const proposalId = receipts
       .find((r) => r.title === "Proposal ID")
       ?.value?.toString();
-    openNewTab(`${explorerProposal}/${proposalId}`);
+    // TOOD: revisit retrieving url (make a proper hook)
+    openNewTab(
+      getNavigationUrl(
+        "proposal_id",
+        explorerLink,
+        proposalId ?? "",
+        lcdEndpoint
+      )
+    );
     onClose?.();
   };